requirements.txt
-i https://pypi.org/simple
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django==3.0.4
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
jinja2==2.11.1
markupsafe==1.1.1
packaging==20.3
psycopg2==2.8.4
pyjwt==1.7.1
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml.clib==0.2.0 ; platform_python_implementation == 'CPython' and python_version < '3.9'
ruamel.yaml==0.16.10
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
